Convert approx_distinct to UDAF
proto style

fix prost

fix clippy

fix doc

fix clippy
